Move some code to ApplicationRepository
