+++ configserver/src/main/java/com/yahoo/vespa/config/server/deploy/Deployment.java
@@ -9,9 +9,6 @@ import com.yahoo.config.provision.DockerImage;
 import com.yahoo.config.provision.HostFilter;
 import com.yahoo.config.provision.Provisioner;
 import java.util.logging.Level;
-import com.yahoo.transaction.NestedTransaction;
-import com.yahoo.transaction.Transaction;
-import com.yahoo.vespa.config.server.ActivationConflictException;
 import com.yahoo.vespa.config.server.ApplicationRepository;
 import com.yahoo.vespa.config.server.ApplicationRepository.ActionTimer;
 import com.yahoo.vespa.config.server.TimeoutBudget;
@@ -145,11 +142,7 @@ public class Deployment implements com.yahoo.config.provision.Deployment {
             CompletionWaiter waiter;
             try (Lock lock = tenant.getApplicationRepo().lock(applicationId)) {
                 previousActiveSession = applicationRepository.getActiveSession(applicationId);
-                waiter = session.createActiveWaiter();
-                NestedTransaction transaction = new NestedTransaction();
-                transaction.add(deactivateCurrentActivateNew(previousActiveSession, session, ignoreSessionStaleFailure));
-                hostProvisioner.ifPresent(provisioner -> provisioner.activate(transaction, applicationId, session.getAllocatedHosts().getHosts()));
-                transaction.commit();
+                waiter = applicationRepository.activate(session, previousActiveSession, applicationId, ignoreSessionStaleFailure);
             }
             catch (RuntimeException e) {
                 throw e;
@@ -186,53 +179,6 @@ public class Deployment implements com.yahoo.config.provision.Deployment {
             throw new IllegalStateException(localSession.logPre() + "Session " + sessionId + " is not prepared");
         } else if (Session.Status.ACTIVATE.equals(localSession.getStatus())) {
             throw new IllegalStateException(localSession.logPre() + "Session " + sessionId + " is already active");
-        }
-    }
-
-    private static Transaction deactivateCurrentActivateNew(Session active, LocalSession prepared, boolean ignoreStaleSessionFailure) {
-        Transaction transaction = prepared.createActivateTransaction();
-        if (isValidSession(active)) {
-            checkIfActiveHasChanged(prepared, active, ignoreStaleSessionFailure);
-            checkIfActiveIsNewerThanSessionToBeActivated(prepared.getSessionId(), active.getSessionId());
-            transaction.add(active.createDeactivateTransaction().operations());
-        }
-        return transaction;
-    }
-
-    private static boolean isValidSession(Session session) {
-        return session != null;
-    }
-
-    private static void checkIfActiveHasChanged(LocalSession session, Session currentActiveSession, boolean ignoreStaleSessionFailure) {
-        long activeSessionAtCreate = session.getActiveSessionAtCreate();
-        log.log(Level.FINE, currentActiveSession.logPre() + "active session id at create time=" + activeSessionAtCreate);
-        if (activeSessionAtCreate == 0) return; // No active session at create
-
-        long sessionId = session.getSessionId();
-        long currentActiveSessionSessionId = currentActiveSession.getSessionId();
-        log.log(Level.FINE, currentActiveSession.logPre() + "sessionId=" + sessionId + 
-                                ", current active session=" + currentActiveSessionSessionId);
-        if (currentActiveSession.isNewerThan(activeSessionAtCreate) &&
-                currentActiveSessionSessionId != sessionId) {
-            String errMsg = currentActiveSession.logPre() + "Cannot activate session " +
-                            sessionId + " because the currently active session (" +
-                            currentActiveSessionSessionId + ") has changed since session " + sessionId +
-                            " was created (was " + activeSessionAtCreate + " at creation time)";
-            if (ignoreStaleSessionFailure) {
-                log.warning(errMsg + " (Continuing because of force.)");
-            } else {
-                throw new ActivationConflictException(errMsg);
-            }
-        }
-    }
-
-    // As of now, config generation is based on session id, and config generation must be a monotonically
-    // increasing number
-    private static void checkIfActiveIsNewerThanSessionToBeActivated(long sessionId, long currentActiveSessionId) {
-        if (sessionId < currentActiveSessionId) {
-            throw new ActivationConflictException("It is not possible to activate session " + sessionId +
-                                          ", because it is older than current active session (" +
-                                          currentActiveSessionId + ")");
         }
     }
