+++ configserver/src/main/java/com/yahoo/vespa/config/server/session/SessionRepository.java
@@ -166,7 +166,7 @@ public class SessionRepository {
                                                               getSessionAppDir(sessionId),
                                                               session.getApplicationPackage(), sessionZooKeeperClient)
                 .getConfigChangeActions();
-        session.setPrepared();
+        setPrepared(session);
         waiter.awaitCompletion(params.getTimeoutBudget().timeLeft());
         return actions;
     }
@@ -489,7 +489,7 @@ public class SessionRepository {
         SessionZooKeeperClient sessionZKClient = createSessionZooKeeperClient(sessionId);
         sessionZKClient.createNewSession(clock.instant());
         Curator.CompletionWaiter waiter = sessionZKClient.getUploadWaiter();
-        LocalSession session = new LocalSession(tenantName, sessionId, applicationPackage, sessionZKClient, applicationRepo);
+        LocalSession session = new LocalSession(tenantName, sessionId, applicationPackage, sessionZKClient);
         waiter.awaitCompletion(timeoutBudget.timeLeft());
         return session;
     }
@@ -547,7 +547,7 @@ public class SessionRepository {
             ApplicationPackage applicationPackage = createApplicationPackage(applicationFile, applicationId,
                                                                              sessionId, currentlyActiveSessionId, false);
             SessionZooKeeperClient sessionZooKeeperClient = createSessionZooKeeperClient(sessionId);
-            return new LocalSession(tenantName, sessionId, applicationPackage, sessionZooKeeperClient, applicationRepo);
+            return new LocalSession(tenantName, sessionId, applicationPackage, sessionZooKeeperClient);
         } catch (Exception e) {
             throw new RuntimeException("Error creating session " + sessionId, e);
         }
@@ -596,7 +596,7 @@ public class SessionRepository {
         File sessionDir = getAndValidateExistingSessionAppDir(sessionId);
         ApplicationPackage applicationPackage = FilesApplicationPackage.fromFile(sessionDir);
         SessionZooKeeperClient sessionZKClient = createSessionZooKeeperClient(sessionId);
-        return new LocalSession(tenantName, sessionId, applicationPackage, sessionZKClient, applicationRepo);
+        return new LocalSession(tenantName, sessionId, applicationPackage, sessionZKClient);
     }
 
     /**
@@ -695,6 +695,20 @@ public class SessionRepository {
 
     private Path lockPath(long sessionId) {
         return locksPath.append(String.valueOf(sessionId));
+    }
+
+    public Transaction createActivateTransaction(Session session) {
+        Transaction transaction = createSetStatusTransaction(session, Session.Status.ACTIVATE);
+        transaction.add(applicationRepo.createPutTransaction(session.sessionZooKeeperClient.readApplicationId(), session.getSessionId()).operations());
+        return transaction;
+    }
+
+    private Transaction createSetStatusTransaction(Session session, Session.Status status) {
+        return session.sessionZooKeeperClient.createWriteStatusTransaction(status);
+    }
+
+    void setPrepared(Session session) {
+        session.setStatus(Session.Status.PREPARE);
     }
 
     private static class FileTransaction extends AbstractTransaction {
